+++ src/features/controls/AdjustInstruments.tsx
