+++ src/features/controls/SongScrubBar.tsx
