+++ src/features/controls/index.ts
@@ -1,2 +1,3 @@
 export { default as AdjustInstruments } from './AdjustInstruments'
 export { default as SongScrubBar } from './SongScrubBar'
+export { VolumeSliderButton } from './VolumeSliderButton'
